Added stack to the CodeGen and a couple comments.
There is a couple comments added to a couple files but the biggest thing
here is the addition of the stack for the offset handling in the
CodeGenerator class. This will be the last checkin before merge changes
back into the master branch.  It will be the baseline before I seriously
start messing with the code to make it work with procedures which is
going to seriously suck.

 On branch p3_conversion
 Changes to be committed:

	modified:   CodeGenerator.C
	modified:   CodeGenerator.H
	modified:   VarRec.H
	modified:   ice9.y