Could fixes for nested if statements.
This doesn't completely work.  There is a problem with jumps at the end
of the if block when they are nested.  Actually it probably happens
otherwise, but I think I can try to push down the jump to next level
down essentially creating a "skip me" before each else if or else so any
code that falls into it will skip it.

 On branch p3_stage_2_reg_fix
 Changes to be committed:

	modified:   ice9.y